Prepack conv weights (#31)
Currently adopt prepacking conv weights in module.to

Pros:
thread-safe
more explicit, compared to implicitly prepack at runtime
Cons:
no input info. meaning that queried format might not be optimal. This is not manifest in conv but this could be a problem for linear weights.
unable to re-pack conv weights if it's been reordered back to plain (maybe even not a con)