Enable BF16 Lazy-Reorder and auto Mix-Precision. (#42)
The user does not need to covert the tensor from float to bfloat16 explicitly to enable the bfloat16 model. IPEX will handle this automatically if the user enables auto mix-precision which is controlled by 'enable_mix_bf16_fp32' flag.
IPEX hides the BF16 data type in DNNL buffer and IPEX always represents the tensor as Float32. If the OP supports BFloat16, IPEX will reorder the buffer to BFloat16 and attach the bf16 buffer to the aten tensor. However, it still is in-place reorder, there might be issues for multi-instances inference.