Wrong shape if the tensor is mixed precision (#100)
If the shape of dil tensor is not as same as its tensor wrapper, then the tensor wrapper will be incorrect because the reorder syncs the shape of dil tensor to its tensor wrapper. In fact, we should only take the dil tensor as the buffer, we should not never sync the meta info of dil tensor to its tensor wrapper. It is JUST a buffer.