Remove the forced queue wait in sync-ed Copy op (#299)
An event wait is already called in sync-ed Copy op.
But sometimes the resouce will be exhausted in
long-term run if there's no queue wait in the
whole run. The queue wait here is to workaround
this issue, to give chance to runtime to recycle
the resource.

For now, we assume the resouce exhaustion issue
has been fixed in runtime, or the recycle strategy
has been improved. Therefore, we are going to
remove the queue wait in this commit.

Signed-off-by: gujinghui <jinghui.gu@intel.com>

Co-authored-by: Kim <kimyangbaochen@vip.qq.com>