Add manual seed for xpu on `test_geometric_kstest` (#1812)
* Add manual seed for xpu

[Note]: This is a very special test case, since when the data type is low accuracy (such as int8), this test could randomly fail when the seed is not set, the low accuracy result will not satisfy the test requirements.

CPU/CUDA by default set this as `1234`, we set it as `12345` respectively. Because xpu has a vectorization accelerate method. This vectorization generated result coincidentally fail at the seed `1234`, thus, we set it to `12345`.

This is because current vectorization does not guarantee that we would get the same result as CUDA. This defect will make some tests potentially fail ( because they could be passed in CUDA env, but fail in xpu case).


Co-authored-by: Jinghui <jinghui.gu@intel.com>