[Profiler] [Kineto+Legacy] Enable Profiler with Legacy & Kineto Support Together (#3400)
* Enable XPU profiler with Kineto support (onetrace version)
 
This is a mixed commit for enabling Kineto + Legacy profiler together
 
In this first version, the base tool is porting as onetrace.
This base tool will be changed to onePTI(temp.) once it was
ready on PoC. Also, this version can't support some features
as:
a. Kineto's memory profiling
b. overhead of PTI tool profiling
c. only enable activities and start trace at anywhen
These features will be fixed once onePTI(temp.) is ready.
 
In this patch, we also fixed some bugs remained in onetrace
to let it worked as great as unitrace. That is, we fixed the bug
which will miss kernels when kernel is small but amount is large.
Also we fixed command list and kernel list's conflicts.
 
We still remain some non-optimized parts here in this first version.
For example, we need to make the profiler start/stop at anytime
and any where instead of a global solution. Also, we will optimize
the serialization operation with protobuf in the future.
 
Signed-off-by: Xunsong, Huang <xunsong.huang@intel.com>