cleanup is_1st_token check (#5507)
Motivation
While the current is_first_token() logic is not implemented correctly, many work arounds like by checking the shapes is used, which is confusing and error prone. We need to fix the is_first_token and replace the workaround to make code clean.

Solution
The root cause is is_1st_token() need to depend on the number of layers(layer_id_static) to get the right timestamp, but reset_layerid() will change layer_id_static and cause this value to be equal 0.
We add another num_of_layers to ensure is_1st_token can get the right value.