Cherry pick vLLM related bug fixes for 2.8 release branch (#5732)
1). FP8 GEMM support input shape [M, B, K] 
keep output shape as [M, B, N].
keep output stride similar as input. In some scenario, the input shape is [M, B, K] and stride is [K, M*K, 1]. We have to make the output stride is [N, M*N, 1] to keep consistency.

2). fix QWEN 32B int4 TP=8 bug
When we run QWEN-32B int4 model with TP=8. The weight will be [80, 5120]. And the group_size of int4 gemm is 128 which is larger then gemm in_feature (80). So, the group_size is changed to 80 which is erroneous.